+++ src/Pages/Dashboard/Admin/AdminDashboard/AdminDashBoard.jsx
@@ -1,43 +1,46 @@
 import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
 import PayMethodModal from '../../Member/PayMethodModal'
-import emailjs from '@emailjs/browser'
-import { toast } from 'react-hot-toast'
+// import emailjs from '@emailjs/browser'
+// import { toast } from 'react-hot-toast'
 import ConfirmationModal from '../../LoanApply/ConfirmationModal'
 import ReminderModal from '../../ReminderModal/ReminderModal'
-// import ConfirmModal from '../../ConfirmModal/ConfirmModal'
 
 const AdminDashboard = () => {
   const [modal,setModal] = useState(false)
   
 
-
-  const sendEmail = () => {
-    const templateParams = {
-      to_name: 'Dibbo Dash',
-      subject: 'Due Payment Remainder!',
-      message: 'Please pay your due with according to table information:',
-      email: 'mdabdurrouf.likhon2@gmail.com',
-      foundation_name: 'Ikhlab Foundation',
-    }
-
-    emailjs
-      .send(
-        'service_55ozcrd',
-        'template_e21lzan',
-        templateParams,
-        'VLd32F4SLKnPSZxsK',
-      )
-      .then(
-        function (response) {
-          console.log('SUCCESS!', response.status, response.text)
-          toast.success('Remainder Send')
-        },
-        function (error) {
-          console.log('FAILED...', error)
-        },
-      )
+  const handleReminder = () => {
+    setModal(true);
+    // sendEmail();
   }
+
+  // const sendEmail = () => {
+  //   const templateParams = {
+  //     to_name: 'Dibbo Dash',
+  //     subject: 'Due Payment Remainder!',
+  //     message: 'Please pay your due with according to table information:',
+  //     email: 'mdabdurrouf.likhon2@gmail.com',
+  //     foundation_name: 'Ikhlab Foundation',
+  //   }
+
+  //   emailjs
+  //     .send(
+  //       'service_55ozcrd',
+  //       'template_e21lzan',
+  //       templateParams,
+  //       'VLd32F4SLKnPSZxsK',
+  //     )
+  //     .then(
+  //       function (response) {
+  //         console.log('SUCCESS!', response.status, response.text)
+  //         toast.success('Remainder Send')
+  //       },
+  //       function (error) {
+  //         console.log('FAILED...', error)
+  //       },
+  //     )
+  // }
 
   const handlePayment = () => {
     console.log('click')
@@ -217,7 +220,8 @@ const AdminDashboard = () => {
               <td className="px-6  text-green-500">01743586381</td>
               <td className="px-6  text-[red]">1000 Tk</td>
               <td className="px-6 ">
-              <button type="button" onClick={() => sendEmail()} className="text-white bg-gradient-to-r from-blue-500 via-blue-600 to-blue-700 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-blue-300 dark:focus:ring-blue-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2">Send Remainder</button>
+              <button type="button" onClick={handleReminder} className="text-white bg-gradient-to-r from-blue-500 via-blue-600 to-blue-700 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-blue-300 dark:focus:ring-blue-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2">Send Remainder</button>
+              {modal && <ReminderModal modal={modal} setModal={setModal} />}
               </td>
             </tr>
             <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
